alto2: replace decoded MIR bit fields
Instead of using a number of UINT8 with the decoded bit fields of the
MIR (micro instruction register), add inline functions to extract the
bits from m_mir. That ought to be faster, because there are fewer memory
accesses in cases where bit fields are not actually used by an instruction.