Imgtool:  Changed the read_sector apparatus to use std::vector
This also eliminated the need for get_sector_size

The asute reviewer will note that a side effect of this change is that in scenarios where we once returned IMGTOOLERR_OUTOFMEMORY, we now throw std::bad_alloc.  This is deliberate.  Having an error code for out of memory conditions no longer makes sense with MAME's embrace of the C++ way of doing things.  It is highly likely that I will follow up this change with one that eliminates IMGTOOLERR_OUTOFMEMORY in favor of throwing std::bad_alloc.

This commit also fixes a recent regression in which we passed nullptr to an std::string ctor