New clone added ------------ Ron Jan  [system11]
This is the original version from 1991, the one already in MAME is Ron Jan Super from 1994. I've renamed the sets accordingly. The Z180 internal ROM hasn't been dumped for this set. The one from Ron Jan Super seems to work fine, but of course the correct ROM should be extracted.