new NOT WORKING  Super Two In One [Edstrom, David Haywood]
this is similar to, but not identical to goldstar.cpp / skylncr.cpp.  it might be possible to merge it in, however my gut feeling is that said driver should be split up into different machine types all using some generic but fleixlbe video device that can handle these 8x8 tiles, 8x32 reel cases as most of them are based on the same basic concepts, just with their own tweaks.

reels, sounds, inputs are not hooked up here.
see
http://forums.bannister.org/ubbthreads.php?ubb=showflat&Number=107569&page=1
for PCB information.