new WORKING   Magnet System (prototype) [David Haywood, Ricky2001, ArcadeHacker, IFW] with the following games Time Scanner (TS 2.0, Magnet System, prototype) Exzisus (EX 1.0, Magnet System, prototype) Xain'd Sleena (SC 3.0, Magnet System, prototype)
this is a checkpoint, there is a sizeable ToDo list but I think this is a good enough state for initial inclusion before working out the remaining issues.  ArcadeHacker has the hardware working for questions etc.
I would especially appreciate help with sound (haven't a clue how it hooks up, tried lots of things, no luck) and some help in going over the CPU comms, especially with the sprite CPU as I think I must be missing something important.  Likewise help from somebody more familiar with the floppy code from MESS so that it can use that, I had no luck in getting it to work with the copy protection or from a rom region, I've kept all floppy related code in a device so it's very easy to swap out so for now what I've got isn't intrusive.

note, very weird idea for a system, they've ported 3 arcade games from other manufacturers, original titles were advertised, but it's likely nothing more of the system exists.
the system uses 5 PCBs, one Z80 on each and instead of having dedicated sprite / tilemap chips they have the Z80s doing software rendering, using 8bpp framebuffers.  2 of the boards are for background planes, 1 of the boards is for sprites.  The hardware has various memory expansions which look like afterthoughts as they're addressed in strange ways.