decapped and read out security pics for umk3 and rmpgwt [Caps0ff]
I've added a new device type and bare bones loading of them so you can see the code running however the communication with the main CPU is not hooked up yet so they still fallback to the simulation code.

also made nbamht a parent, because the nbamht / nbahangt situation really isn't too different to the umk3 / mk3 or sf2ce / sf2 situations, it's a distinct new release.