Preliminary attempt to fix regression caused by poor understanding of get_default_card_software() (#2122)
Tafoid brought to my attention a recent regression caused by my softlist changes.  Specifically, this failed:

    mame odyssey2 -cart1 voice -cart2 bees

The reason is that the recent refactoring caused some critical softlist logic to be bypassed.  Essentially device_slot_interface::get_default_card_software() provides a technique for devices to inspect softlists and select slot options to be specified.  In the case above, the selection of the "voice" software list item in the first cartridge slot requires get_default_card_software() to be called to specify the o2_voice slot device, which in turn adds the second cartridge slot.  Skipping this caused '-cart2' to stop being recognized.

This is an attempt to fix the problem.  In some ways, this doesn't go far enough - the astute observer will note that mame_options::reevaluate_slot_options() looks a lot like mame_options::update_slot_options(); part of me wants to just rip the latter out.  In the long run, device_slot_interface::get_default_card_software() should be reevaluated; having a hook is a good idea but it should be a pure function that doesn't alter state.