Fixed regression when loading multipart softlists
Pernod found a regression introduced in the 0.183 softlist refactoring whereby multi-part softlist items would not distribute to multiple slots.  The problem was that the old code was relying on the image slots being loaded into the core.  This is not the way the new system works, so I've added a hook into software_list_device::find_mountable_image() that allows the new approach to work.