[CoCo 3]  Fixed an issue clearing interrupts on the GIME implementation
The normal way to acknowledge a GIME interrupt is by reading from $FF92/3.  However, the act of disabling interrupts by writing to the same registers will also do the trick, and we were not emulating this technique.

Thanks to Glen Hewlett and Tim Lindner for helping to identify this problem