Changes to make get_default_card_software() less stupid
The goals with this change is to make get_default_card_software() a bit more standalone by making it a const method that does not mutate the state of the device_image_interface.  This is done by passing in a small structure that encapsulates the minimum of information that get_default_card_software() needs.

This also eliminates the need for device_image_interface::open_image_file()

I agree with Sarayan that get_default_card_software() is terrible and needs to ultimately go away.  This is a small step in that direction.

Lastly, I don't care for the name of get_default_card_software_hook (or get_default_card_software() for that matter).  If anybody has better ideas, let me know.