Fixed out of bounds issues in NeoGeo memory access
The NeoGeo driver exposes a number of different memory regions as vectors.  Accessing the pointers was done through a '&vector[0]' pattern.  This caused problems when the region was size zero; I changed the code to return null pointers in these scenarios.

This was reported by Robbbert in response to recent regressions (the command line was 'mame aes bjourney'), but it seems to be present in MAME 0.184.