Goodbye MACHINE_CONFIG_FRAGMENT, it was nice knowing you.
(nw) This is a pretty minimal change.  The point where the root device is added has been moved
from the MACHINE_CONFIG_START macro to the constructor of the machine configuration class (made
possible by giving drivers their own device types).  This isn't the final change in this area.
The root device is still being handled specially in that its configuration comes from the game
driver structure.  This needs to be harmonised with regular devices.  But that's a job for
another day.