new working clones  Treasure Island (DECO Cassette) (Region D) [David Haywood]
this is the ctisland3 set that has been in MAME since forever.  compared to the other ctisland sets it uses a different encryption and has a region code of 'D' in the header.
we haven't seen any 'D' region roms before (A is Japan, B is assumed to be US) so I've had to handcraft a bios rom for it (the only difference between the A/B bios roms is the single byte it checks in the header anyway)