m62 proposed cleanups.
I've split the massive PROM blocks up into something more modern looking (the single PROM region with all PROMs crammed into it is an old MAME thing from when only one PROM region was allowed)  I've also made use of multipe palettes and gfxdecode structures (required using rgb32 to draw to)

As the boards are quite modular this seems to make sense to me, the palette for the sprites and the sprite decode + sprite palette decode is the same across all games while the rest of the hardware changes a bit.  This should allow us to step closer to converting the various boards into devices and piecing them together more logically.