* Get past fatal error on selecting software part from internal UI * Sort software part menu in internal UI (was ordered by hash)
This preserves the option priority behaviour from 0.187 when setting the
instance option to the desired part.

Note that software part selection from internal UI doesn't actually
work.  For example with zorba:cpm, no matter which part you select,
you'll always end up with flop1 in A: and flop2 in B:.  This actually
did work properly in 0.187, so something has gone wrong in between.