gameboy: fix SGB VRAM transfers (#2504)
* gameboy.xml: remove misleading comment

Sachen 4B-003 was recently added

Signed-off-by: Tauwasser <tauwasser@tauwasser.eu>

* gameboy: fix Super Game Boy VRAM transfer

A basic implementation of VRAM transfer. It fixes a number of games and removes
the SGB border hack. However, it's very likely that the bahvior is much more
complex. The old implementation was good enough for the majority of games,
so this should suffice until such time when SGB is implemented on top of SNES.

The attribute data was resized to 4096 bytes, so a whole VRAM transfer can take place
even though only 4050 bytes are used. The idea is that the whole 4096 bytes are
_probably_ transferred to WRAM and a game might theoretically upload a small executable
and use that data. However, running native SNES code is currently unsupported anyway.

Signed-off-by: Tauwasser <tauwasser@tauwasser.eu>

* gameboy: various code style/comment fixes

 - return GB_MBC_NONE instead of magic 0 value
 - add MLT_REQ case in sgb code and mention where it's actually handled
 - add PAL_PRI to list of known SGB commands (not implemented)
 - fix two comments

Signed-off-by: Tauwasser <tauwasser@tauwasser.eu>

* gameboy: coding style fixes for gb_lcd

Signed-off-by: Tauwasser <tauwasser@tauwasser.eu>