Clean up Galeco Dallas hookup device.
There is no need for a second device for World Rally and Glass, it's the
same thing.  How the address lines are hooked up for host memory access
is part of the driver and can be done with a mirror/mask in the address
map provided by the driver.  There's also no need for the SRAM
trampolines in the device, that can be done entirely through address
maps.

The drivers need trampolines to do the address swizzling.  We should be
able to get rid of these when we get better support for 8-bit views of
16-bit components (dooyong drivers also have ugly trampolines for doing
this kind of thing).