ui: improve layout of art/info box, support mousewheel scroll in info text, allow home/end to jump to beginning/end of filter list, consolidate logic
(nw) A number of vestigial constants have been removed.  Some hacky
input types that were just used as a trick to pass information between
menu functions are gone.  MACHINE_NO_STANDALONE is a relic from when
drivers were used as arbitrary ROM containers, and in a world of
first-class devices this is no longer necessary.