floppy: Add drive LED callback
This currently only works if you set the drive select lines correctly,
which the upd765 does now. Hook it up to the psi98 driver and use the
TEAC FD-55F drive instead of the generic one.

This is some basic first work to better model the real hardware and
floppy drive setups in the future.