Replaced hacky Kaypro keyboard that bypassed the SIO with emulated Kaypro 10 keyboard. [Vas Crabb, rfka01, TeamEurope]
(nw) This exposes other issues in MAME.  Some incredibly poorly-written
code in the keyboard MCU depends on the fact that the T register is
updated during S4 then copied to A during S4 when mov a,t is executed.
If it doesn't see zero in this register when mov a,t is executed
immediately before the timer interrupt is taken, it hangs.
Really, the MCS-48 core needs to be re-written so it works at S-cycle
level (if not clock cycle leve), but for now I've worked around this
case in the least intrusive way possible.
This also exposes issues in the Z80DART code.  The keyboard communicates
0.8% slower than the nominal 300 Baud.  This works perfectly in real
life, but it causes occasional corrupted characters with MAME's
implementation, particularly if you hold a key down and let it repeat.
The Z80DART device needs to be fixed so it samples closer to the middle
of the bit intervals and re-aquires on each start bit.
Finally, I haven't hooked up the Kaypro's serial port Baud rate
genrators, or exposed the serial ports as slots.  I'll leave that for
someone else to worry about.