fixing/improving metadata on CoCo FDCs: CP-450 and CD-6809
Thanks to Daniel Campos & Datassette.org for the CP-450 disk rom. MAME was calling it CP400_FDC, but the correct name of the disk controller unit is "Prológica CP-450". The ROM was correct (verified checksum), but I renamed it to fix the previous naming mistake. Unfortunately this FDC unit is rare, so I still don't know whats the IC reference for the EPROM on the PCB.

There is a photo of the CP-450 disk controller unit at:
https://datassette.org/softwares/tandy-trs-color/cp-450-basic-disco-v-10
http://files.datassette.org/softwares/img/wp_20141212_22_08_26_pro.jpg