Fixes to make basic boot on HP9000/300 (#3466)
* hp9k_3xx: add dummy gpib handler

BASIC expects that there is an internal gpib card present.
It crashes without, so let's add some dummy handlers for now.

Signed-off-by: Sven Schnelle <svens@stackframe.org>

* hp9k_3xx: add fallback buserror handler

HP BASIC and the bootrom rely on buserror to figure out
hardware configuration. Adding a global buserror handler
makes BASIC boot on models >= 340. 310 doesn't show a picture
with that handler, so let's leave them out for now.

Signed-off-by: Sven Schnelle <svens@stackframe.org>

* hp9k_3xx: remove unneeded buserror handlers

Now that we have the fallback buserror handler in place,
we don't need the other buserror handlers anymore.

Signed-off-by: Sven Schnelle <svens@stackframe.org>

* hp9k_3xx: log accesses to gpib registers

Signed-off-by: Sven Schnelle <svens@stackframe.org>