PowerVR2: implement color
PowerVR2 represents color using two components: a base color and an optional
offset color.  The base color is combined with the incoming texture sample (if
any) according to the fixed-function specified by the tsinstruction in the
per-polygon parameters, and then the offset is added to this color (with the
exception of alpha; the offset color has an alpha component but it does not
appear to ever be used).

The base color and offset color can be specified on a per-vertex basis (as
either a group of four 8-bit integers or a group of four single-precision
floating points) or they can be specified on a per-polygon basis (in
floating-point only) and then scaled based on per-vertex scalar values (also
floating point).