mc146818: support direct-access bus hookup method. [R. Belmont]
This chip has 8 address/data lines and an address/data strobe.  Most
users hooked it up so the low-order address bit is the address/data
strobe, allowing you to write an address and then read/write data
at an adjacent address.

But if you don't mind using a few more gates you can latch the
low-order address bits from the CPU into the chip then latch the
data in or out of it so it works just like a RAM.  The MIPS DECstations
did this in their I/O ASICs.