current hng64 / tlcs870 notes, tweaks etc.
I'll probably tear out the simulation at some point soon so that we're trying to hook this up live, IMHO it's not far off being able to do something useful, just need to find the various triggers and make things like the timer timing realistic.  Don't think the serial stuff is going to matter for the basic hookup.  None of the games are marked as working yet anyway, so hopefully won't be a problem.