rx2030: wip checkpoint (nw)
Most hardware is in place now, initial diagnostics pass and it's possible to enter both the boot monitor and load/run the sash (stand-alone shell) from disk. The keyboard no longer works in sash, possibly because it's expecting a MIPS-specific keyboard to be connected. RISC/os loads apparently successfully but dies shortly after, probably due to lack of TLB in r3000.cpp.