new clones - U.N. Defense Force: Earth Joker (US / Japan, set 2) [Phil Bennett]
note, it's very close to the parent (71 differences in each ROM) but there are enough intentional changes (including a small block of 00/ff right before the data ends) that it can't be considered a direct replacement for the bad set; these aren't all single bit changes that could be attibuted to the bitrot.

also changed descriptions, as these sets aren't Japan only except for the prototype.
also lowered sound levels so that they don't clip badly due to the MSM