pcat101: connect matrix and ports
(nw) After this change, the keyboard sends scan codes to the host successfully, but fails to receive/process commands from the host. Failure is probably due to differences between the 6805U3 that's really in the keyboard, and the 68705U3 that's the best match in mame at the moment, so requires expanded 6805 emulation.