Fix kidniki schematics after review.
The kungfu master schematics differ from the kidnik schematics in
wiring D4, D5 and Q4 with other components. The manual corrections
to the kungfu master schematics make sense and thus are used here.

opamp XU1.B according to schematics has no feedback loop between
output and inputs. The arrangement of components in the schematic
however indicate that this is not the case and indeed a connection
exists. This results in sounds at output XU1.14 to contain more
detail.