tutankham: Implement starfield, better colors. [Guru, Couriersud]
Used color code from galaxian for better color approximation.
The starfield code was added to better document the 084 custom chip.
There are two star fields implemented now. The active one can be
selected in Machine Configuration.

a) Scramble Starfield

1/3 star width like scramble/galaxian, starfield code same as scramble.
This one is similar to the Konami starfield video below. The 084 uses
the same signals as scramble.

b) Bootleg Starfield

Guru provided schematics of a daughter-board which on a Konami bootleg
replaced the custom chip 084. This starfield matches the starfield
(picture) of this bootleg.
At least there is now a starfield which matches a real-life(tm) bootleg.

None of the two alterntives reproduces the star field which can be
observed in the Stern videos on youtube.

The Konami video mentioned in video/tutankhm.cpp is different from
tutankhm. Looks like we are missing an early board dump