Sanity prevailed (nw)
There are some bugs, probably linked to the sample counter (I thought
it was a frame counter for a start...).  But at least the code is not
aneurysm-provoking anymore.

One thing to do would be to turn the k573dio into a
device_mixer_interface to have the connection to the speakers done at
the driver device level.  Volunteers?