fmtowns.cpp: fix crash regression in System Sacom games (#5397) [r09]
This fixes a regression introduced with commit b9ee9c7c2860c2745a1fdb096ce1c75792080e81 that causes semi-random emulator crashes (at least in non-debug builds) in 38-man Kilo no Kokuu and Yami no Ketsuzoku Special. Apparently they are trying to set colors for palette indexes higher than 15 in 16-color mode, and the m_entry_color array goes out of bounds.

I'm not sure if the hardware discards everything but the 4 least significant bits of the palette index, or if it ignores the color change completely, but it doesn't seem to make any difference in practice, so I'm assuming the former.