Update PCB information in chinagat.cpp
Current document says it's unknown where are the background images
stored. But it should be in the TRJ-100 custom package, that needs a
custom adapter to dump it.

Also add a side note that the TRJ-100 chip generates 5MHz clock for
some unknown reasons, and used by some ICs as a clock. This may help
to solve the speed problems in the future.