mc6845: address cursor display edge case in interlaced mode.
It appears that the cursor is not displayed when the start raster
address is equal to the maximum raster address in the 'interlace and
video' mode. At least this addresses a claimed regression in the
'apricot' machine. To be explore further on hardware.

The maximum raster address was also off by one in these calculations.