tv990: cursor fix
The addresses of the rows in the vram do not increase monotonically
with the screen scan lines and the cursor address appears to follow
these addresses so the cursor row can not be computed from the
address. Rather compute the cursor column from the difference of the
cursor address offset and the address off of the start of the row, and
implicitly match the rows when comparing the columns. This might not be
perfect, but it is a big improvement and no issues have been spotted.