Opengl render driver: fix bug preventing adjustment of ... [Couriersud]
gamma, brightness and contrast. Starting MAME with defaults,
i.e. gamma == 1.0, no rgb palette will be allocated.
If gamma than is changed, a new palette will be allocated. But the
texture will not be updated because the palette is not checked.
This fix will check for palette changes.
Note: if the palette is changed inplace, i.e. without reallocation, this
will fail. The correct fix would be that the texture palette is not only
a pointer to rgb_t * but to texture_palette_type. texture_palette_type
would hold the pointer, the length and a sequential id which is checked
in addition to pointer equality.