jaguar.cpp: Memory-related refactoring (nw) - Set width and endianness for regions, including cartridges - Make init code safer and more efficient - De-duplicate some memory maps - Separate state class for jaguarcd
I've tried to keep this driver from regressing any further than it already has while cleaning up the code and adjusting region usage to comply with the soon-to-be-committed "truth in endianness" requirement. It's possible that some regions should really be 16 bits instead of 32 or vice versa, and/or that DSP endianness may be incorrect.