Revert "8255: missed one driver (nw)"
This reverts commit 1631d2d2b1ef026f7bd75c8126642a2d50d5df48.

Revert "8255: set default tristate to 0 (nw)"

This reverts commit b891c77972386e044ac1778de87fa5e1f0d44788.

This has been shown to break bw2, and it's very likely it breaks other
computers in subtle ways or in ways that are only noticeable when you
use a certain peripheral.  It's far too dangerous, particularly late in
a development cycle and with Christmas/new year coming.  You can't push
a change like this without checking every use of the device.

Please fix any drivers you subsequently modified to depend on the change.