mips1: revert unnecessary/unwanted initialization (nw)
All of these variables are either:
1. Already initialized in the constructor or start/reset; or
2. Guarded against read-before-write by program logic; or
3. Documented as being in an undefined state at device power-up.

Regarding the last point, my view is that if the documentation states the value is unpredictable/undefined at power up, I'd rather not initialize it to an arbitrary value. The absence of initialization in the code counts as documenting this behaviour, or conversely, initializing it to an arbitrary zero documents something which is not factual. While there is an argument for consistency in emulation, I'd rather expose and address any such undocumented hardware behaviour.

I believe the only required and missing initialization here is for integer register zero, which was indeed a bug. If there are any other test cases which fail after this reversion, please let me know and I'll investigate and address them.