Add Multitech MPF-II as clone of tk2000
MPF-II is one of the earliest Apple II clone but not totally compatible
with.

It does not have a text mode, The address of the second graphic buffer
is different as well as the address of the keyboard input.
It have only one extension slot whit somme little pinning differences,
no joystick interface, a printer and a keyboard connector.

tk2000 drivers seems to be far more close to MPF-II, with the same
memory mapping, ram/rom banking in upper 16K, compatible keyboard
interface, same apple2 like video interface.

This drivers partially tested and seems fine with MPF-II rom.
Tested : Video, keyboard, sound.